+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGED:
 - update gradle to new version 7.4
 - update of test-dependency crypt-api to new version 7.7
 - update of test-dependency crypt-data to new version 7.11.1
+- replaced test-dependency 'de.alpharogroup:auth-security' with new 'io.github.astrapi69:auth' and 'io.github.astrapi69:auth-api' in major version 6
 
 
 Version 5.1